requirements.txt
dataclasses-json>=0.5.7
typing-extensions>=4.0.0
cerebras-cloud-sdk>=1.0.0
python-dotenv>=1.0.0
requests>-2.31.0